+++ aiowinrm/errors.py
@@ -1,2 +1,17 @@
+
+
 class AIOWinRMException(Exception):
     pass
+
+
+class SoapException(Exception):
+    pass
+
+
+class SoapTimeout(SoapException):
+    pass
+
+
+class WsManException(Exception):
+    pass
+
