--- aiowinrm/tmp.py
@@ -1,29 +1,0 @@
-from aiowinrm.psrp.fragmenter import Fragmenter
-from aiowinrm.psrp.messages import create_session_capability_message, init_runspace_pool_message
-from aiowinrm.soap.protocol import create_power_shell_payload
-import uuid
-
-import lxml.etree as etree
-
-# print(py == ru)
-
-py = [0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 3, 0, 0, 0, 224, 2, 0, 0, 0, 2, 0, 1, 0, 36, 20, 41, 170, 150, 174, 59, 67, 138, 17, 225, 240, 4, 52, 187, 118, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 239, 187, 191, 60, 79, 98, 106, 32, 82, 101, 102, 73, 100, 61, 34, 48, 34, 62, 10, 32, 32, 60, 77, 83, 62, 10, 32, 32, 32, 32, 60, 86, 101, 114, 115, 105, 111, 110, 32, 78, 61, 34, 112, 114, 111, 116, 111, 99, 111, 108, 118, 101, 114, 115, 105, 111, 110, 34, 62, 50, 46, 51, 60, 47, 86, 101, 114, 115, 105, 111, 110, 62, 10, 32, 32, 32, 32, 60, 86, 101, 114, 115, 105, 111, 110, 32, 78, 61, 34, 80, 83, 86, 101, 114, 115, 105, 111, 110, 34, 62, 50, 46, 48, 60, 47, 86, 101, 114, 115, 105, 111, 110, 62, 10, 32, 32, 32, 32, 60, 86, 101, 114, 115, 105, 111, 110, 32, 78, 61, 34, 83, 101, 114, 105, 97, 108, 105, 122, 97, 116, 105, 111, 110, 86, 101, 114, 115, 105, 111, 110, 34, 62, 49, 46, 49, 46, 48, 46, 49, 60, 47, 86, 101, 114, 115, 105, 111, 110, 62, 10, 32, 32, 60, 47, 77, 83, 62, 10, 60, 47, 79, 98, 106, 62, 0, 0, 0, 0, 0, 0, 0, 2, 0, 0, 0, 0, 0, 0, 0, 0, 3, 0, 0, 29, 20, 2, 0, 0, 0, 4, 0, 1, 0, 36, 20, 41, 170, 150, 174, 59, 67, 138, 17, 225, 240, 4, 52, 187, 118, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 239, 187, 191, 60, 79, 98, 106, 32, 82, 101, 102, 73, 100, 61, 34, 48, 34, 62, 10, 32, 32, 60, 77, 83, 62, 10, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 77, 105, 110, 82, 117, 110, 115, 112, 97, 99, 101, 115, 34, 62, 49, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 77, 97, 120, 82, 117, 110, 115, 112, 97, 99, 101, 115, 34, 62, 49, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 80, 83, 84, 104, 114, 101, 97, 100, 79, 112, 116, 105, 111, 110, 115, 34, 32, 82, 101, 102, 73, 100, 61, 34, 49, 34, 62, 10, 32, 32, 32, 32, 32, 32, 60, 84, 78, 32, 82, 101, 102, 73, 100, 61, 34, 48, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 60, 84, 62, 83, 121, 115, 116, 101, 109, 46, 77, 97, 110, 97, 103, 101, 109, 101, 110, 116, 46, 65, 117, 116, 111, 109, 97, 116, 105, 111, 110, 46, 82, 117, 110, 115, 112, 97, 99, 101, 115, 46, 80, 83, 84, 104, 114, 101, 97, 100, 79, 112, 116, 105, 111, 110, 115, 60, 47, 84, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 60, 84, 62, 83, 121, 115, 116, 101, 109, 46, 69, 110, 117, 109, 60, 47, 84, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 60, 84, 62, 83, 121, 115, 116, 101, 109, 46, 86, 97, 108, 117, 101, 84, 121, 112, 101, 60, 47, 84, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 60, 84, 62, 83, 121, 115, 116, 101, 109, 46, 79, 98, 106, 101, 99, 116, 60, 47, 84, 62, 10, 32, 32, 32, 32, 32, 32, 60, 47, 84, 78, 62, 10, 32, 32, 32, 32, 32, 32, 60, 84, 111, 83, 116, 114, 105, 110, 103, 62, 68, 101, 102, 97, 117, 108, 116, 60, 47, 84, 111, 83, 116, 114, 105, 110, 103, 62, 10, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 62, 48, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 65, 112, 97, 114, 116, 109, 101, 110, 116, 83, 116, 97, 116, 101, 34, 32, 82, 101, 102, 73, 100, 61, 34, 50, 34, 62, 10, 32, 32, 32, 32, 32, 32, 60, 84, 78, 32, 82, 101, 102, 73, 100, 61, 34, 49, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 60, 84, 62, 83, 121, 115, 116, 101, 109, 46, 84, 104, 114, 101, 97, 100, 105, 110, 103, 46, 65, 112, 97, 114, 116, 109, 101, 110, 116, 83, 116, 97, 116, 101, 60, 47, 84, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 60, 84, 62, 83, 121, 115, 116, 101, 109, 46, 69, 110, 117, 109, 60, 47, 84, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 60, 84, 62, 83, 121, 115, 116, 101, 109, 46, 86, 97, 108, 117, 101, 84, 121, 112, 101, 60, 47, 84, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 60, 84, 62, 83, 121, 115, 116, 101, 109, 46, 79, 98, 106, 101, 99, 116, 60, 47, 84, 62, 10, 32, 32, 32, 32, 32, 32, 60, 47, 84, 78, 62, 10, 32, 32, 32, 32, 32, 32, 60, 84, 111, 83, 116, 114, 105, 110, 103, 62, 85, 110, 107, 110, 111, 119, 110, 60, 47, 84, 111, 83, 116, 114, 105, 110, 103, 62, 10, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 62, 50, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 65, 112, 112, 108, 105, 99, 97, 116, 105, 111, 110, 65, 114, 103, 117, 109, 101, 110, 116, 115, 34, 32, 82, 101, 102, 73, 100, 61, 34, 51, 34, 62, 10, 32, 32, 32, 32, 32, 32, 60, 84, 78, 32, 82, 101, 102, 73, 100, 61, 34, 50, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 60, 84, 62, 83, 121, 115, 116, 101, 109, 46, 77, 97, 110, 97, 103, 101, 109, 101, 110, 116, 46, 65, 117, 116, 111, 109, 97, 116, 105, 111, 110, 46, 80, 83, 80, 114, 105, 109, 105, 116, 105, 118, 101, 68, 105, 99, 116, 105, 111, 110, 97, 114, 121, 60, 47, 84, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 60, 84, 62, 83, 121, 115, 116, 101, 109, 46, 67, 111, 108, 108, 101, 99, 116, 105, 111, 110, 115, 46, 72, 97, 115, 104, 116, 97, 98, 108, 101, 60, 47, 84, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 60, 84, 62, 83, 121, 115, 116, 101, 109, 46, 79, 98, 106, 101, 99, 116, 60, 47, 84, 62, 10, 32, 32, 32, 32, 32, 32, 60, 47, 84, 78, 62, 10, 32, 32, 32, 32, 32, 32, 60, 68, 67, 84, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 60, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 83, 32, 78, 61, 34, 75, 101, 121, 34, 62, 80, 83, 86, 101, 114, 115, 105, 111, 110, 84, 97, 98, 108, 101, 60, 47, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 86, 97, 108, 117, 101, 34, 32, 82, 101, 102, 73, 100, 61, 34, 52, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 84, 78, 82, 101, 102, 32, 82, 101, 102, 73, 100, 61, 34, 50, 34, 32, 47, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 68, 67, 84, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 83, 32, 78, 61, 34, 75, 101, 121, 34, 62, 80, 83, 86, 101, 114, 115, 105, 111, 110, 60, 47, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 86, 101, 114, 115, 105, 111, 110, 32, 78, 61, 34, 86, 97, 108, 117, 101, 34, 62, 53, 46, 48, 46, 49, 49, 48, 56, 50, 46, 49, 48, 48, 48, 60, 47, 86, 101, 114, 115, 105, 111, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 83, 32, 78, 61, 34, 75, 101, 121, 34, 62, 80, 83, 67, 111, 109, 112, 97, 116, 105, 98, 108, 101, 86, 101, 114, 115, 105, 111, 110, 115, 60, 47, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 86, 97, 108, 117, 101, 34, 32, 82, 101, 102, 73, 100, 61, 34, 53, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 84, 78, 32, 82, 101, 102, 73, 100, 61, 34, 51, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 84, 62, 83, 121, 115, 116, 101, 109, 46, 86, 101, 114, 115, 105, 111, 110, 91, 93, 60, 47, 84, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 84, 62, 83, 121, 115, 116, 101, 109, 46, 65, 114, 114, 97, 121, 60, 47, 84, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 84, 62, 83, 121, 115, 116, 101, 109, 46, 79, 98, 106, 101, 99, 116, 60, 47, 84, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 84, 78, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 76, 83, 84, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 86, 101, 114, 115, 105, 111, 110, 62, 49, 46, 48, 60, 47, 86, 101, 114, 115, 105, 111, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 86, 101, 114, 115, 105, 111, 110, 62, 50, 46, 48, 60, 47, 86, 101, 114, 115, 105, 111, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 86, 101, 114, 115, 105, 111, 110, 62, 51, 46, 48, 60, 47, 86, 101, 114, 115, 105, 111, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 86, 101, 114, 115, 105, 111, 110, 62, 52, 46, 48, 60, 47, 86, 101, 114, 115, 105, 111, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 86, 101, 114, 115, 105, 111, 110, 62, 53, 46, 48, 46, 49, 49, 48, 56, 50, 46, 49, 48, 48, 48, 60, 47, 86, 101, 114, 115, 105, 111, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 76, 83, 84, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 83, 32, 78, 61, 34, 75, 101, 121, 34, 62, 67, 76, 82, 86, 101, 114, 115, 105, 111, 110, 60, 47, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 86, 101, 114, 115, 105, 111, 110, 32, 78, 61, 34, 86, 97, 108, 117, 101, 34, 62, 52, 46, 48, 46, 51, 48, 51, 49, 57, 46, 52, 50, 48, 48, 48, 60, 47, 86, 101, 114, 115, 105, 111, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 83, 32, 78, 61, 34, 75, 101, 121, 34, 62, 66, 117, 105, 108, 100, 86, 101, 114, 115, 105, 111, 110, 60, 47, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 86, 101, 114, 115, 105, 111, 110, 32, 78, 61, 34, 86, 97, 108, 117, 101, 34, 62, 49, 48, 46, 48, 46, 49, 49, 48, 56, 50, 46, 49, 48, 48, 48, 60, 47, 86, 101, 114, 115, 105, 111, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 83, 32, 78, 61, 34, 75, 101, 121, 34, 62, 87, 83, 77, 97, 110, 83, 116, 97, 99, 107, 86, 101, 114, 115, 105, 111, 110, 60, 47, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 86, 101, 114, 115, 105, 111, 110, 32, 78, 61, 34, 86, 97, 108, 117, 101, 34, 62, 51, 46, 48, 60, 47, 86, 101, 114, 115, 105, 111, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 83, 32, 78, 61, 34, 75, 101, 121, 34, 62, 80, 83, 82, 101, 109, 111, 116, 105, 110, 103, 80, 114, 111, 116, 111, 99, 111, 108, 86, 101, 114, 115, 105, 111, 110, 60, 47, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 86, 101, 114, 115, 105, 111, 110, 32, 78, 61, 34, 86, 97, 108, 117, 101, 34, 62, 50, 46, 51, 60, 47, 86, 101, 114, 115, 105, 111, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 83, 32, 78, 61, 34, 75, 101, 121, 34, 62, 83, 101, 114, 105, 97, 108, 105, 122, 97, 116, 105, 111, 110, 86, 101, 114, 115, 105, 111, 110, 60, 47, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 86, 101, 114, 115, 105, 111, 110, 32, 78, 61, 34, 86, 97, 108, 117, 101, 34, 62, 49, 46, 49, 46, 48, 46, 49, 60, 47, 86, 101, 114, 115, 105, 111, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 68, 67, 84, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 60, 47, 68, 67, 84, 62, 10, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 72, 111, 115, 116, 73, 110, 102, 111, 34, 32, 82, 101, 102, 73, 100, 61, 34, 54, 34, 62, 10, 32, 32, 32, 32, 32, 32, 60, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 95, 104, 111, 115, 116, 68, 101, 102, 97, 117, 108, 116, 68, 97, 116, 97, 34, 32, 82, 101, 102, 73, 100, 61, 34, 55, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 100, 97, 116, 97, 34, 32, 82, 101, 102, 73, 100, 61, 34, 56, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 84, 78, 32, 82, 101, 102, 73, 100, 61, 34, 52, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 84, 62, 83, 121, 115, 116, 101, 109, 46, 67, 111, 108, 108, 101, 99, 116, 105, 111, 110, 115, 46, 72, 97, 115, 104, 116, 97, 98, 108, 101, 60, 47, 84, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 84, 62, 83, 121, 115, 116, 101, 109, 46, 79, 98, 106, 101, 99, 116, 60, 47, 84, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 84, 78, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 68, 67, 84, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 75, 101, 121, 34, 62, 57, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 86, 97, 108, 117, 101, 34, 32, 82, 101, 102, 73, 100, 61, 34, 57, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 83, 32, 78, 61, 34, 84, 34, 62, 83, 121, 115, 116, 101, 109, 46, 83, 116, 114, 105, 110, 103, 60, 47, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 83, 32, 78, 61, 34, 86, 34, 62, 67, 58, 92, 100, 101, 118, 92, 107, 105, 116, 99, 104, 101, 110, 45, 118, 97, 103, 114, 97, 110, 116, 60, 47, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 75, 101, 121, 34, 62, 56, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 86, 97, 108, 117, 101, 34, 32, 82, 101, 102, 73, 100, 61, 34, 49, 48, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 83, 32, 78, 61, 34, 84, 34, 62, 83, 121, 115, 116, 101, 109, 46, 77, 97, 110, 97, 103, 101, 109, 101, 110, 116, 46, 65, 117, 116, 111, 109, 97, 116, 105, 111, 110, 46, 72, 111, 115, 116, 46, 83, 105, 122, 101, 60, 47, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 86, 34, 32, 82, 101, 102, 73, 100, 61, 34, 49, 49, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 119, 105, 100, 116, 104, 34, 62, 49, 57, 57, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 104, 101, 105, 103, 104, 116, 34, 62, 53, 50, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 75, 101, 121, 34, 62, 55, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 86, 97, 108, 117, 101, 34, 32, 82, 101, 102, 73, 100, 61, 34, 49, 50, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 83, 32, 78, 61, 34, 84, 34, 62, 83, 121, 115, 116, 101, 109, 46, 77, 97, 110, 97, 103, 101, 109, 101, 110, 116, 46, 65, 117, 116, 111, 109, 97, 116, 105, 111, 110, 46, 72, 111, 115, 116, 46, 83, 105, 122, 101, 60, 47, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 86, 34, 32, 82, 101, 102, 73, 100, 61, 34, 49, 51, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 119, 105, 100, 116, 104, 34, 62, 56, 48, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 104, 101, 105, 103, 104, 116, 34, 62, 53, 50, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 75, 101, 121, 34, 62, 54, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 86, 97, 108, 117, 101, 34, 32, 82, 101, 102, 73, 100, 61, 34, 49, 52, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 83, 32, 78, 61, 34, 84, 34, 62, 83, 121, 115, 116, 101, 109, 46, 77, 97, 110, 97, 103, 101, 109, 101, 110, 116, 46, 65, 117, 116, 111, 109, 97, 116, 105, 111, 110, 46, 72, 111, 115, 116, 46, 83, 105, 122, 101, 60, 47, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 86, 34, 32, 82, 101, 102, 73, 100, 61, 34, 49, 53, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 119, 105, 100, 116, 104, 34, 62, 56, 48, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 104, 101, 105, 103, 104, 116, 34, 62, 50, 53, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 75, 101, 121, 34, 62, 53, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 86, 97, 108, 117, 101, 34, 32, 82, 101, 102, 73, 100, 61, 34, 49, 54, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 83, 32, 78, 61, 34, 84, 34, 62, 83, 121, 115, 116, 101, 109, 46, 77, 97, 110, 97, 103, 101, 109, 101, 110, 116, 46, 65, 117, 116, 111, 109, 97, 116, 105, 111, 110, 46, 72, 111, 115, 116, 46, 83, 105, 122, 101, 60, 47, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 86, 34, 32, 82, 101, 102, 73, 100, 61, 34, 49, 55, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 119, 105, 100, 116, 104, 34, 62, 50, 53, 48, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 104, 101, 105, 103, 104, 116, 34, 62, 57, 57, 57, 57, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 75, 101, 121, 34, 62, 52, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 86, 97, 108, 117, 101, 34, 32, 82, 101, 102, 73, 100, 61, 34, 49, 56, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 83, 32, 78, 61, 34, 84, 34, 62, 83, 121, 115, 116, 101, 109, 46, 73, 110, 116, 51, 50, 60, 47, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 86, 34, 62, 50, 53, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 75, 101, 121, 34, 62, 51, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 86, 97, 108, 117, 101, 34, 32, 82, 101, 102, 73, 100, 61, 34, 49, 57, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 83, 32, 78, 61, 34, 84, 34, 62, 83, 121, 115, 116, 101, 109, 46, 77, 97, 110, 97, 103, 101, 109, 101, 110, 116, 46, 65, 117, 116, 111, 109, 97, 116, 105, 111, 110, 46, 72, 111, 115, 116, 46, 67, 111, 111, 114, 100, 105, 110, 97, 116, 101, 115, 60, 47, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 86, 34, 32, 82, 101, 102, 73, 100, 61, 34, 50, 48, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 120, 34, 62, 48, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 121, 34, 62, 57, 57, 55, 52, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 75, 101, 121, 34, 62, 50, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 86, 97, 108, 117, 101, 34, 32, 82, 101, 102, 73, 100, 61, 34, 50, 49, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 83, 32, 78, 61, 34, 84, 34, 62, 83, 121, 115, 116, 101, 109, 46, 77, 97, 110, 97, 103, 101, 109, 101, 110, 116, 46, 65, 117, 116, 111, 109, 97, 116, 105, 111, 110, 46, 72, 111, 115, 116, 46, 67, 111, 111, 114, 100, 105, 110, 97, 116, 101, 115, 60, 47, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 86, 34, 32, 82, 101, 102, 73, 100, 61, 34, 50, 50, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 120, 34, 62, 48, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 121, 34, 62, 57, 57, 57, 56, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 75, 101, 121, 34, 62, 49, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 86, 97, 108, 117, 101, 34, 32, 82, 101, 102, 73, 100, 61, 34, 50, 51, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 83, 32, 78, 61, 34, 84, 34, 62, 83, 121, 115, 116, 101, 109, 46, 67, 111, 110, 115, 111, 108, 101, 67, 111, 108, 111, 114, 60, 47, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 86, 34, 62, 48, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 75, 101, 121, 34, 62, 48, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 86, 97, 108, 117, 101, 34, 32, 82, 101, 102, 73, 100, 61, 34, 50, 52, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 83, 32, 78, 61, 34, 84, 34, 62, 83, 121, 115, 116, 101, 109, 46, 67, 111, 110, 115, 111, 108, 101, 67, 111, 108, 111, 114, 60, 47, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 86, 34, 62, 55, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 68, 67, 84, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 60, 66, 32, 78, 61, 34, 95, 105, 115, 72, 111, 115, 116, 78, 117, 108, 108, 34, 62, 102, 97, 108, 115, 101, 60, 47, 66, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 60, 66, 32, 78, 61, 34, 95, 105, 115, 72, 111, 115, 116, 85, 73, 78, 117, 108, 108, 34, 62, 102, 97, 108, 115, 101, 60, 47, 66, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 60, 66, 32, 78, 61, 34, 95, 105, 115, 72, 111, 115, 116, 82, 97, 119, 85, 73, 78, 117, 108, 108, 34, 62, 102, 97, 108, 115, 101, 60, 47, 66, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 60, 66, 32, 78, 61, 34, 95, 117, 115, 101, 82, 117, 110, 115, 112, 97, 99, 101, 72, 111, 115, 116, 34, 62, 102, 97, 108, 115, 101, 60, 47, 66, 62, 10, 32, 32, 32, 32, 32, 32, 60, 47, 77, 83, 62, 10, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 60, 47, 77, 83, 62, 10, 60, 47, 79, 98, 106, 62]
-ru = [0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 3, 0, 0, 0, 225, 2, 0, 0, 0, 2, 0, 1, 0, 36, 20, 41, 170, 150, 174, 59, 67, 138, 17, 225, 240, 4, 52, 187, 118, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 239, 187, 191, 60, 79, 98, 106, 32, 82, 101, 102, 73, 100, 61, 34, 48, 34, 62, 10, 32, 32, 60, 77, 83, 62, 10, 32, 32, 32, 32, 60, 86, 101, 114, 115, 105, 111, 110, 32, 78, 61, 34, 112, 114, 111, 116, 111, 99, 111, 108, 118, 101, 114, 115, 105, 111, 110, 34, 62, 50, 46, 51, 60, 47, 86, 101, 114, 115, 105, 111, 110, 62, 10, 32, 32, 32, 32, 60, 86, 101, 114, 115, 105, 111, 110, 32, 78, 61, 34, 80, 83, 86, 101, 114, 115, 105, 111, 110, 34, 62, 50, 46, 48, 60, 47, 86, 101, 114, 115, 105, 111, 110, 62, 10, 32, 32, 32, 32, 60, 86, 101, 114, 115, 105, 111, 110, 32, 78, 61, 34, 83, 101, 114, 105, 97, 108, 105, 122, 97, 116, 105, 111, 110, 86, 101, 114, 115, 105, 111, 110, 34, 62, 49, 46, 49, 46, 48, 46, 49, 60, 47, 86, 101, 114, 115, 105, 111, 110, 62, 10, 32, 32, 60, 47, 77, 83, 62, 10, 60, 47, 79, 98, 106, 62, 10, 0, 0, 0, 0, 0, 0, 0, 2, 0, 0, 0, 0, 0, 0, 0, 0, 3, 0, 0, 29, 21, 2, 0, 0, 0, 4, 0, 1, 0, 36, 20, 41, 170, 150, 174, 59, 67, 138, 17, 225, 240, 4, 52, 187, 118, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 239, 187, 191, 60, 79, 98, 106, 32, 82, 101, 102, 73, 100, 61, 34, 48, 34, 62, 10, 32, 32, 60, 77, 83, 62, 10, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 77, 105, 110, 82, 117, 110, 115, 112, 97, 99, 101, 115, 34, 62, 49, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 77, 97, 120, 82, 117, 110, 115, 112, 97, 99, 101, 115, 34, 62, 49, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 80, 83, 84, 104, 114, 101, 97, 100, 79, 112, 116, 105, 111, 110, 115, 34, 32, 82, 101, 102, 73, 100, 61, 34, 49, 34, 62, 10, 32, 32, 32, 32, 32, 32, 60, 84, 78, 32, 82, 101, 102, 73, 100, 61, 34, 48, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 60, 84, 62, 83, 121, 115, 116, 101, 109, 46, 77, 97, 110, 97, 103, 101, 109, 101, 110, 116, 46, 65, 117, 116, 111, 109, 97, 116, 105, 111, 110, 46, 82, 117, 110, 115, 112, 97, 99, 101, 115, 46, 80, 83, 84, 104, 114, 101, 97, 100, 79, 112, 116, 105, 111, 110, 115, 60, 47, 84, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 60, 84, 62, 83, 121, 115, 116, 101, 109, 46, 69, 110, 117, 109, 60, 47, 84, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 60, 84, 62, 83, 121, 115, 116, 101, 109, 46, 86, 97, 108, 117, 101, 84, 121, 112, 101, 60, 47, 84, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 60, 84, 62, 83, 121, 115, 116, 101, 109, 46, 79, 98, 106, 101, 99, 116, 60, 47, 84, 62, 10, 32, 32, 32, 32, 32, 32, 60, 47, 84, 78, 62, 10, 32, 32, 32, 32, 32, 32, 60, 84, 111, 83, 116, 114, 105, 110, 103, 62, 68, 101, 102, 97, 117, 108, 116, 60, 47, 84, 111, 83, 116, 114, 105, 110, 103, 62, 10, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 62, 48, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 65, 112, 97, 114, 116, 109, 101, 110, 116, 83, 116, 97, 116, 101, 34, 32, 82, 101, 102, 73, 100, 61, 34, 50, 34, 62, 10, 32, 32, 32, 32, 32, 32, 60, 84, 78, 32, 82, 101, 102, 73, 100, 61, 34, 49, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 60, 84, 62, 83, 121, 115, 116, 101, 109, 46, 84, 104, 114, 101, 97, 100, 105, 110, 103, 46, 65, 112, 97, 114, 116, 109, 101, 110, 116, 83, 116, 97, 116, 101, 60, 47, 84, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 60, 84, 62, 83, 121, 115, 116, 101, 109, 46, 69, 110, 117, 109, 60, 47, 84, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 60, 84, 62, 83, 121, 115, 116, 101, 109, 46, 86, 97, 108, 117, 101, 84, 121, 112, 101, 60, 47, 84, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 60, 84, 62, 83, 121, 115, 116, 101, 109, 46, 79, 98, 106, 101, 99, 116, 60, 47, 84, 62, 10, 32, 32, 32, 32, 32, 32, 60, 47, 84, 78, 62, 10, 32, 32, 32, 32, 32, 32, 60, 84, 111, 83, 116, 114, 105, 110, 103, 62, 85, 110, 107, 110, 111, 119, 110, 60, 47, 84, 111, 83, 116, 114, 105, 110, 103, 62, 10, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 62, 50, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 65, 112, 112, 108, 105, 99, 97, 116, 105, 111, 110, 65, 114, 103, 117, 109, 101, 110, 116, 115, 34, 32, 82, 101, 102, 73, 100, 61, 34, 51, 34, 62, 10, 32, 32, 32, 32, 32, 32, 60, 84, 78, 32, 82, 101, 102, 73, 100, 61, 34, 50, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 60, 84, 62, 83, 121, 115, 116, 101, 109, 46, 77, 97, 110, 97, 103, 101, 109, 101, 110, 116, 46, 65, 117, 116, 111, 109, 97, 116, 105, 111, 110, 46, 80, 83, 80, 114, 105, 109, 105, 116, 105, 118, 101, 68, 105, 99, 116, 105, 111, 110, 97, 114, 121, 60, 47, 84, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 60, 84, 62, 83, 121, 115, 116, 101, 109, 46, 67, 111, 108, 108, 101, 99, 116, 105, 111, 110, 115, 46, 72, 97, 115, 104, 116, 97, 98, 108, 101, 60, 47, 84, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 60, 84, 62, 83, 121, 115, 116, 101, 109, 46, 79, 98, 106, 101, 99, 116, 60, 47, 84, 62, 10, 32, 32, 32, 32, 32, 32, 60, 47, 84, 78, 62, 10, 32, 32, 32, 32, 32, 32, 60, 68, 67, 84, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 60, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 83, 32, 78, 61, 34, 75, 101, 121, 34, 62, 80, 83, 86, 101, 114, 115, 105, 111, 110, 84, 97, 98, 108, 101, 60, 47, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 86, 97, 108, 117, 101, 34, 32, 82, 101, 102, 73, 100, 61, 34, 52, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 84, 78, 82, 101, 102, 32, 82, 101, 102, 73, 100, 61, 34, 50, 34, 32, 47, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 68, 67, 84, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 83, 32, 78, 61, 34, 75, 101, 121, 34, 62, 80, 83, 86, 101, 114, 115, 105, 111, 110, 60, 47, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 86, 101, 114, 115, 105, 111, 110, 32, 78, 61, 34, 86, 97, 108, 117, 101, 34, 62, 53, 46, 48, 46, 49, 49, 48, 56, 50, 46, 49, 48, 48, 48, 60, 47, 86, 101, 114, 115, 105, 111, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 83, 32, 78, 61, 34, 75, 101, 121, 34, 62, 80, 83, 67, 111, 109, 112, 97, 116, 105, 98, 108, 101, 86, 101, 114, 115, 105, 111, 110, 115, 60, 47, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 86, 97, 108, 117, 101, 34, 32, 82, 101, 102, 73, 100, 61, 34, 53, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 84, 78, 32, 82, 101, 102, 73, 100, 61, 34, 51, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 84, 62, 83, 121, 115, 116, 101, 109, 46, 86, 101, 114, 115, 105, 111, 110, 91, 93, 60, 47, 84, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 84, 62, 83, 121, 115, 116, 101, 109, 46, 65, 114, 114, 97, 121, 60, 47, 84, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 84, 62, 83, 121, 115, 116, 101, 109, 46, 79, 98, 106, 101, 99, 116, 60, 47, 84, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 84, 78, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 76, 83, 84, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 86, 101, 114, 115, 105, 111, 110, 62, 49, 46, 48, 60, 47, 86, 101, 114, 115, 105, 111, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 86, 101, 114, 115, 105, 111, 110, 62, 50, 46, 48, 60, 47, 86, 101, 114, 115, 105, 111, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 86, 101, 114, 115, 105, 111, 110, 62, 51, 46, 48, 60, 47, 86, 101, 114, 115, 105, 111, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 86, 101, 114, 115, 105, 111, 110, 62, 52, 46, 48, 60, 47, 86, 101, 114, 115, 105, 111, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 86, 101, 114, 115, 105, 111, 110, 62, 53, 46, 48, 46, 49, 49, 48, 56, 50, 46, 49, 48, 48, 48, 60, 47, 86, 101, 114, 115, 105, 111, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 76, 83, 84, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 83, 32, 78, 61, 34, 75, 101, 121, 34, 62, 67, 76, 82, 86, 101, 114, 115, 105, 111, 110, 60, 47, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 86, 101, 114, 115, 105, 111, 110, 32, 78, 61, 34, 86, 97, 108, 117, 101, 34, 62, 52, 46, 48, 46, 51, 48, 51, 49, 57, 46, 52, 50, 48, 48, 48, 60, 47, 86, 101, 114, 115, 105, 111, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 83, 32, 78, 61, 34, 75, 101, 121, 34, 62, 66, 117, 105, 108, 100, 86, 101, 114, 115, 105, 111, 110, 60, 47, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 86, 101, 114, 115, 105, 111, 110, 32, 78, 61, 34, 86, 97, 108, 117, 101, 34, 62, 49, 48, 46, 48, 46, 49, 49, 48, 56, 50, 46, 49, 48, 48, 48, 60, 47, 86, 101, 114, 115, 105, 111, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 83, 32, 78, 61, 34, 75, 101, 121, 34, 62, 87, 83, 77, 97, 110, 83, 116, 97, 99, 107, 86, 101, 114, 115, 105, 111, 110, 60, 47, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 86, 101, 114, 115, 105, 111, 110, 32, 78, 61, 34, 86, 97, 108, 117, 101, 34, 62, 51, 46, 48, 60, 47, 86, 101, 114, 115, 105, 111, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 83, 32, 78, 61, 34, 75, 101, 121, 34, 62, 80, 83, 82, 101, 109, 111, 116, 105, 110, 103, 80, 114, 111, 116, 111, 99, 111, 108, 86, 101, 114, 115, 105, 111, 110, 60, 47, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 86, 101, 114, 115, 105, 111, 110, 32, 78, 61, 34, 86, 97, 108, 117, 101, 34, 62, 50, 46, 51, 60, 47, 86, 101, 114, 115, 105, 111, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 83, 32, 78, 61, 34, 75, 101, 121, 34, 62, 83, 101, 114, 105, 97, 108, 105, 122, 97, 116, 105, 111, 110, 86, 101, 114, 115, 105, 111, 110, 60, 47, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 86, 101, 114, 115, 105, 111, 110, 32, 78, 61, 34, 86, 97, 108, 117, 101, 34, 62, 49, 46, 49, 46, 48, 46, 49, 60, 47, 86, 101, 114, 115, 105, 111, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 68, 67, 84, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 60, 47, 68, 67, 84, 62, 10, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 72, 111, 115, 116, 73, 110, 102, 111, 34, 32, 82, 101, 102, 73, 100, 61, 34, 54, 34, 62, 10, 32, 32, 32, 32, 32, 32, 60, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 95, 104, 111, 115, 116, 68, 101, 102, 97, 117, 108, 116, 68, 97, 116, 97, 34, 32, 82, 101, 102, 73, 100, 61, 34, 55, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 100, 97, 116, 97, 34, 32, 82, 101, 102, 73, 100, 61, 34, 56, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 84, 78, 32, 82, 101, 102, 73, 100, 61, 34, 52, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 84, 62, 83, 121, 115, 116, 101, 109, 46, 67, 111, 108, 108, 101, 99, 116, 105, 111, 110, 115, 46, 72, 97, 115, 104, 116, 97, 98, 108, 101, 60, 47, 84, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 84, 62, 83, 121, 115, 116, 101, 109, 46, 79, 98, 106, 101, 99, 116, 60, 47, 84, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 84, 78, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 68, 67, 84, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 75, 101, 121, 34, 62, 57, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 86, 97, 108, 117, 101, 34, 32, 82, 101, 102, 73, 100, 61, 34, 57, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 83, 32, 78, 61, 34, 84, 34, 62, 83, 121, 115, 116, 101, 109, 46, 83, 116, 114, 105, 110, 103, 60, 47, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 83, 32, 78, 61, 34, 86, 34, 62, 67, 58, 92, 100, 101, 118, 92, 107, 105, 116, 99, 104, 101, 110, 45, 118, 97, 103, 114, 97, 110, 116, 60, 47, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 75, 101, 121, 34, 62, 56, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 86, 97, 108, 117, 101, 34, 32, 82, 101, 102, 73, 100, 61, 34, 49, 48, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 83, 32, 78, 61, 34, 84, 34, 62, 83, 121, 115, 116, 101, 109, 46, 77, 97, 110, 97, 103, 101, 109, 101, 110, 116, 46, 65, 117, 116, 111, 109, 97, 116, 105, 111, 110, 46, 72, 111, 115, 116, 46, 83, 105, 122, 101, 60, 47, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 86, 34, 32, 82, 101, 102, 73, 100, 61, 34, 49, 49, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 119, 105, 100, 116, 104, 34, 62, 49, 57, 57, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 104, 101, 105, 103, 104, 116, 34, 62, 53, 50, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 75, 101, 121, 34, 62, 55, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 86, 97, 108, 117, 101, 34, 32, 82, 101, 102, 73, 100, 61, 34, 49, 50, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 83, 32, 78, 61, 34, 84, 34, 62, 83, 121, 115, 116, 101, 109, 46, 77, 97, 110, 97, 103, 101, 109, 101, 110, 116, 46, 65, 117, 116, 111, 109, 97, 116, 105, 111, 110, 46, 72, 111, 115, 116, 46, 83, 105, 122, 101, 60, 47, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 86, 34, 32, 82, 101, 102, 73, 100, 61, 34, 49, 51, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 119, 105, 100, 116, 104, 34, 62, 56, 48, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 104, 101, 105, 103, 104, 116, 34, 62, 53, 50, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 75, 101, 121, 34, 62, 54, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 86, 97, 108, 117, 101, 34, 32, 82, 101, 102, 73, 100, 61, 34, 49, 52, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 83, 32, 78, 61, 34, 84, 34, 62, 83, 121, 115, 116, 101, 109, 46, 77, 97, 110, 97, 103, 101, 109, 101, 110, 116, 46, 65, 117, 116, 111, 109, 97, 116, 105, 111, 110, 46, 72, 111, 115, 116, 46, 83, 105, 122, 101, 60, 47, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 86, 34, 32, 82, 101, 102, 73, 100, 61, 34, 49, 53, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 119, 105, 100, 116, 104, 34, 62, 56, 48, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 104, 101, 105, 103, 104, 116, 34, 62, 50, 53, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 75, 101, 121, 34, 62, 53, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 86, 97, 108, 117, 101, 34, 32, 82, 101, 102, 73, 100, 61, 34, 49, 54, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 83, 32, 78, 61, 34, 84, 34, 62, 83, 121, 115, 116, 101, 109, 46, 77, 97, 110, 97, 103, 101, 109, 101, 110, 116, 46, 65, 117, 116, 111, 109, 97, 116, 105, 111, 110, 46, 72, 111, 115, 116, 46, 83, 105, 122, 101, 60, 47, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 86, 34, 32, 82, 101, 102, 73, 100, 61, 34, 49, 55, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 119, 105, 100, 116, 104, 34, 62, 50, 53, 48, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 104, 101, 105, 103, 104, 116, 34, 62, 57, 57, 57, 57, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 75, 101, 121, 34, 62, 52, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 86, 97, 108, 117, 101, 34, 32, 82, 101, 102, 73, 100, 61, 34, 49, 56, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 83, 32, 78, 61, 34, 84, 34, 62, 83, 121, 115, 116, 101, 109, 46, 73, 110, 116, 51, 50, 60, 47, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 86, 34, 62, 50, 53, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 75, 101, 121, 34, 62, 51, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 86, 97, 108, 117, 101, 34, 32, 82, 101, 102, 73, 100, 61, 34, 49, 57, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 83, 32, 78, 61, 34, 84, 34, 62, 83, 121, 115, 116, 101, 109, 46, 77, 97, 110, 97, 103, 101, 109, 101, 110, 116, 46, 65, 117, 116, 111, 109, 97, 116, 105, 111, 110, 46, 72, 111, 115, 116, 46, 67, 111, 111, 114, 100, 105, 110, 97, 116, 101, 115, 60, 47, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 86, 34, 32, 82, 101, 102, 73, 100, 61, 34, 50, 48, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 120, 34, 62, 48, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 121, 34, 62, 57, 57, 55, 52, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 75, 101, 121, 34, 62, 50, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 86, 97, 108, 117, 101, 34, 32, 82, 101, 102, 73, 100, 61, 34, 50, 49, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 83, 32, 78, 61, 34, 84, 34, 62, 83, 121, 115, 116, 101, 109, 46, 77, 97, 110, 97, 103, 101, 109, 101, 110, 116, 46, 65, 117, 116, 111, 109, 97, 116, 105, 111, 110, 46, 72, 111, 115, 116, 46, 67, 111, 111, 114, 100, 105, 110, 97, 116, 101, 115, 60, 47, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 86, 34, 32, 82, 101, 102, 73, 100, 61, 34, 50, 50, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 120, 34, 62, 48, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 121, 34, 62, 57, 57, 57, 56, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 75, 101, 121, 34, 62, 49, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 86, 97, 108, 117, 101, 34, 32, 82, 101, 102, 73, 100, 61, 34, 50, 51, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 83, 32, 78, 61, 34, 84, 34, 62, 83, 121, 115, 116, 101, 109, 46, 67, 111, 110, 115, 111, 108, 101, 67, 111, 108, 111, 114, 60, 47, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 86, 34, 62, 48, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 75, 101, 121, 34, 62, 48, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 79, 98, 106, 32, 78, 61, 34, 86, 97, 108, 117, 101, 34, 32, 82, 101, 102, 73, 100, 61, 34, 50, 52, 34, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 83, 32, 78, 61, 34, 84, 34, 62, 83, 121, 115, 116, 101, 109, 46, 67, 111, 110, 115, 111, 108, 101, 67, 111, 108, 111, 114, 60, 47, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 73, 51, 50, 32, 78, 61, 34, 86, 34, 62, 55, 60, 47, 73, 51, 50, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 69, 110, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 68, 67, 84, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 77, 83, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 60, 66, 32, 78, 61, 34, 95, 105, 115, 72, 111, 115, 116, 78, 117, 108, 108, 34, 62, 102, 97, 108, 115, 101, 60, 47, 66, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 60, 66, 32, 78, 61, 34, 95, 105, 115, 72, 111, 115, 116, 85, 73, 78, 117, 108, 108, 34, 62, 102, 97, 108, 115, 101, 60, 47, 66, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 60, 66, 32, 78, 61, 34, 95, 105, 115, 72, 111, 115, 116, 82, 97, 119, 85, 73, 78, 117, 108, 108, 34, 62, 102, 97, 108, 115, 101, 60, 47, 66, 62, 10, 32, 32, 32, 32, 32, 32, 32, 32, 60, 66, 32, 78, 61, 34, 95, 117, 115, 101, 82, 117, 110, 115, 112, 97, 99, 101, 72, 111, 115, 116, 34, 62, 102, 97, 108, 115, 101, 60, 47, 66, 62, 10, 32, 32, 32, 32, 32, 32, 60, 47, 77, 83, 62, 10, 32, 32, 32, 32, 60, 47, 79, 98, 106, 62, 10, 32, 32, 60, 47, 77, 83, 62, 10, 60, 47, 79, 98, 106, 62, 10]
-
-print([chr(c) for c in py])
-print([chr(c) for c in ru])
-
-session_id = uuid.UUID('AA291424-AE96-433B-8A11-E1F00434BB76')
-cap_msg = create_session_capability_message(session_id)
-runspace_msg = init_runspace_pool_message(session_id)
-
-
-creation_payload = Fragmenter.messages_to_fragment_bytes((cap_msg, runspace_msg))
-
-print([c for c in creation_payload])
-
-payload = etree.tostring(create_power_shell_payload(session_id, creation_payload))
-
-print(payload.decode('utf-8'))
-
-
